requirements.txt
streamlit==1.24.1
firebase-admin==6.0.1
python-dotenv==1.0.0
openai==0.27.8
anthropic==0.48.0